Update logic.py
need to create a folder named "channel_images" in the same directory as the logic.py file. The images should be named "channel_1.png", "channel_2.png", and "channel_3.png" 



I was thinking we can use netflix, hulu, disney plus, etc. for images when the channels are changed or we can use actual TV channels.